pyproject.toml
[tool.poetry]
name = "test-task-sockets"
version = "0.1.0"
description = ""
authors = ["Andrey Isaev <an.isaev7@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_sockets"}]

[tool.poetry.dependencies]
python = "^3.10"
eventlet = "^0.33.3"
python-socketio = "^5.8.0"
faker = "^19.1.0"
cryptography = "^41.0.2"
pydantic = "^2.0.3"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
